extract most of event list logic to a component
